% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpca.R
\name{do.robpca}
\alias{do.robpca}
\title{Perform robust PCA (Hubert PCA)}
\usage{
do.robpca(data, ncp = 10, ...)
}
\arguments{
\item{data}{Log-transformed gene expression counts}
}
\value{
A list containing the output of PcaHubert, the (default) PC scores, and the PC loadings
}
\description{
This function performs the variant of robust PCA developed by Mia Hubert et al. on a
matrix of counts
}
