% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpca.R
\name{RPCAPlot}
\alias{RPCAPlot}
\title{Plot the results of rPCA with cells colored by their identity class.}
\usage{
RPCAPlot(object, ...)
}
\arguments{
\item{object}{Seurat object}

\item{...}{Additional parameters to DimPlot; for example, which dimensions to
plot.}
}
\description{
Plot the results of rPCA with cells colored by their identity class.
}
