% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpca.R
\name{RunRobPCA}
\alias{RunRobPCA}
\title{Perform robust PCA (Hubert PCA) with modified PC scores on a Seurat object}
\usage{
RunRobPCA(object, npcs = 10, pc.genes = NULL,
  use.modified.pcscores = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{npcs}{Number of principal components to calculate}

\item{pc.genes}{Genes as input to PC. If \code{pc.genes==NULL}, the var.genes
slot is used. If var.genes is empty and \code{pc.genes==NULL}, then all
genes are used.}

\item{use.modified.pcscores}{If \code{FALSE}, then the raw pc score output
from robust PCA is used. If \code{TRUE}, then pc scores are replaced with
the sum of the top 30 genes by positive loading minus the sum of the top 30
genes by negative loading. Each gene is scaled to a max of 1 and min of 0.}
}
\value{
A Seurat object with the 'rpca' field filled.
}
\description{
\code{scRobustPCA} performs the variant of robust PCA developed by Mia Hubert
et al. on the gene expression matrix "data" in a Seurat object. Set
reduction.type='rpca' in other Seurat functions to use the rPCA results, for
example to calculate clusters with FindClusters.
}
\examples{

}
